+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - New Topic button moved
 - Boards now display their topics in a child LiveView (instead of LiveComponent)
+- Deleting posts is now confirmed with Phoenix.HTML's data-confirm prompt
 
 ## [0.8.1] - 2021-10-08